+++ .eslintrc.cjs
@@ -1,0 +1,25 @@
+module.exports = {
+    root: true,
+    env: { browser: true, es2020: true, node: true },
+    extends: [
+        'eslint:recommended',
+        'plugin:react/recommended',
+        'plugin:react/jsx-runtime',
+        'plugin:react-hooks/recommended',
+    ],
+    ignorePatterns: ['dist', '.eslintrc.cjs'],
+    parserOptions: { ecmaVersion: 'latest', sourceType: 'module' },
+    settings: { react: { version: '18.2' } },
+    plugins: ['react-refresh'],
+    rules: {
+        'react-refresh/only-export-components': [
+            'warn',
+            { allowConstantExport: true },
+        ],
+        'max-lines': ['error', { max: 250, skipBlankLines: true, skipComments: true }],
+        'complexity': ['warn', 15],
+        'max-params': ['warn', 4],
+        'no-unused-vars': ['warn', { argsIgnorePattern: '^_' }],
+        'no-console': ['warn', { allow: ['warn', 'error'] }],
+    },
+}